requirements.txt
streamlit==1.32.0
pandas==2.2.1
pillow==10.2.0
python-dotenv==1.0.1 